Added new object vector
All game objects should be derived form Object. A vector of unique_ptrs includes all objects in the game. It is iterated every frame. Object is derived from transformable and drawable. Has an IsDrawable boolean to determine if it should be drawn every frame. This should be set upon construction and be class specific.